Refactor user search
The change refactors the user search feature by moving offset and limit verifications to the list of retrieved users instead of the database query. This approach allows reducing the number of expensive database operations, thereby optimizing performance. The refactoring also eliminates the UserSearchDTO structure which is found to be redundant, making the codebase cleaner and more efficient.