Refactor subscription handling and reduce context deadline
This commit simplifies subscription handling by removing the JwtManager from the NewSubscriptionHandler function and renaming various entities related to following and followers to a more generic SubscriptionUserDTO. The deadline for context in the HandleGetSubscriptions was also reduced from 1000 seconds to 10 seconds for better performance. A new constant 'SubscriptionTypeParamKey' was added in url_keys.go for improved code clarity.